functions/load_csv/requirements.txt
PyYAML
google-cloud-bigquery
google-cloud-storage
flask
